chg: move startComparison function to lib/
